add mwe phrases
MWE - multi word expression.
They usually describe some entity - full name of a person, name of an
organization or a city.
It doesn't make sense to use parts of MWE as independent phrases.
So we detect MWE at first, and then add to them modifiers from the
sentence or use MWE as modifiers for some words.