phrases: add phrase types; generalize MWEs
Before this commit MWEs were stored differently than main phrases.
It was inconvenient especially for serializing.
This commit ensures that phrases and MWEs are stored in one place,
specifically in Sent._phrases.
To distinguish between MWE and ordinary phrase PhraseType was introduced
that is assigned to each phrase in the builder.